Redirect storage to experiments/ subfolder if experiment exists
In many cases, it is useful to start and execution from an existing experiment
with different configuration. For instance, restarting from a certain
checkpoint with different configuration. To handle this use case,
machinable will now detect if a given storage URL points to an
existing experiment and rewrite the storage URL to the
experiments/ subfolder (closes #15)