Disable automatic checkpoint handling and allow user specific settings
Before this change, the component was attempting to load a checkpoint as
soon as the CHECKPOINT flag was set. However, there are many scenarios
in which the checkoint handling needs to happen at particular times
(e.g. before model initialisation). To account for these cases, we let
the user handle any checkpoint related logic for which config can be passed
in via the checkpoint argument/flag. machinable will not implement any
checkpoint related logic but rather focus on managing checkpoint related
configuration and logging (closes #13)