Vendor improvements
Adds an option to directly append machinable.yaml files. Unlike the existing vendor imports API, this allows to integrate models without renames and can thus be useful to modularise larger chunks of the machinable.yaml into subfiles. Along with it, this introduces a new registration event to intercept the vendor resolution as well as direct support for git repository vendors.