New submission API
Reading of data from storage was previously handled by the ExperimentStorage abstraction. This required multiple objects and methods that were not available at the top import level. It also overloaded the Experiment and Storage terms. This breaking change removes the various APIs in favor of a single Submission interface that is available from the top import level. In future, the submission concept might enable easier in-memory manipulation of execution parameters as it is not sematically restricted to be storage.