Remove enderscore config
This removes an experimental feature where any config value ending in an underscore (example_) would be ignored during look up. The issue with this feature is that it can lead to unexpected behavior when explicitly setting an update in the version ({"example_": True}). It is not obvious what the behavior should be in such case, so in conclusion is probably better to simply remove such a feature.